Initial storage implementation
This relies on localStorage, but webExtensions do not have
access to that. Instead they have access to a priviledged
web extensions storag which I don't know much about.

As a follow up to this I should abstract the parts where I talk
the the underlying storage system, so that we can easily swap
that out.